Crate port_sniffer

port_sniffer_cli

A simple asynchronous TCP port scanner in Rust with a progress bar. Scans a range of ports on a target IP concurrently using Tokio.

Example Usage

port_sniffer_cli --ip 192.168.0.1 --start_port 1 --end_port 1024 --concurrency 50

Constants

ABOUT 🔒

About description

APP_NAME 🔒

Application name

AUTHOR 🔒

Author

CHANNEL_BUFFER_SIZE 🔒

Buffer size for the mpsc channel

DEFAULT_CONCURRENCY 🔒

Default concurrency value

DEFAULT_END_PORT 🔒

Default end port

DEFAULT_START_PORT 🔒

Default start port

HELP_CONCURRENCY 🔒

Help message for concurrency argument

HELP_IP 🔒

Help message for IP argument

LONG_CONCURRENCY 🔒

Long name for concurrency argument

LONG_END_PORT 🔒

Long name for end port

LONG_IP 🔒

Long name for IP argument

LONG_START_PORT 🔒

Long name for start port

MAX_PORT 🔒

Maximum valid TCP port

MIN_PORT 🔒

Minimum valid TCP port

SHORT_CONCURRENCY 🔒

Short name for concurrency argument

SHORT_END_PORT 🔒

Short name for end port

SHORT_START_PORT 🔒

Short name for start port

VERSION 🔒

Version

Functions

main 🔒

Main asynchronous entry point using Tokio runtime

scan 🔒

Attempts to connect to a given IP and port asynchronously.