port_sniffer
Function scan

pub(crate) async fn scan(
    tx: Sender<u16>,
    port: u16,
    addr: IpAddr,
    pb: Arc<ProgressBar>,
)

Attempts to connect to a given IP and port asynchronously.

If the connection succeeds, the port is sent through the mpsc channel and the progress bar is incremented.

Arguments

tx - Channel sender to report open ports

port - Port number to test

addr - Target IP address

pb - Shared progress bar